+++ react-app/src/components/EditPostsComponent/EditPostModal.js
@@ -1,0 +1,14 @@
+import { useState } from "react";
+
+function EditPostModal() {
+    const [showModal, setShowModal] = useState(false);
+
+    return (
+        <>
+            <button onClick={() => setShowModal(true)}>
+                Edit
+            </button>
+            
+        </>
+    )
+}
